refactor: 命名規則を統一
