+++ src/main/kotlin/dev/usbharu/hideout/repository/JwtRefreshTokenRepository.kt
@@ -2,7 +2,7 @@ package dev.usbharu.hideout.repository
 
 import dev.usbharu.hideout.domain.model.hideout.entity.JwtRefreshToken
 
-interface IJwtRefreshTokenRepository {
+interface JwtRefreshTokenRepository {
     suspend fun generateId(): Long
 
     suspend fun save(token: JwtRefreshToken)
