+++ src/main/kotlin/dev/usbharu/hideout/repository/PostRepository.kt
@@ -3,7 +3,7 @@ package dev.usbharu.hideout.repository
 import dev.usbharu.hideout.domain.model.hideout.entity.Post
 
 @Suppress("LongParameterList")
-interface IPostRepository {
+interface PostRepository {
     suspend fun generateId(): Long
     suspend fun save(post: Post): Post
     suspend fun delete(id: Long)
