+++ src/main/kotlin/dev/usbharu/hideout/service/core/MetaService.kt
@@ -3,7 +3,7 @@ package dev.usbharu.hideout.service.core
 import dev.usbharu.hideout.domain.model.hideout.entity.Jwt
 import dev.usbharu.hideout.domain.model.hideout.entity.Meta
 
-interface IMetaService {
+interface MetaService {
     suspend fun getMeta(): Meta
     suspend fun updateMeta(meta: Meta)
     suspend fun getJwtMeta(): Jwt
