+++ src/main/kotlin/dev/usbharu/hideout/service/post/PostService.kt
@@ -3,6 +3,6 @@ package dev.usbharu.hideout.service.post
 import dev.usbharu.hideout.domain.model.hideout.dto.PostCreateDto
 import dev.usbharu.hideout.domain.model.hideout.entity.Post
 
-interface IPostService {
+interface PostService {
     suspend fun createLocal(post: PostCreateDto): Post
 }
