+++ src/main/kotlin/dev/usbharu/hideout/service/reaction/ReactionService.kt
@@ -1,6 +1,6 @@
 package dev.usbharu.hideout.service.reaction
 
-interface IReactionService {
+interface ReactionService {
     suspend fun receiveReaction(name: String, domain: String, userId: Long, postId: Long)
     suspend fun sendReaction(name: String, userId: Long, postId: Long)
     suspend fun removeReaction(userId: Long, postId: Long)
