+++ src/main/kotlin/dev/usbharu/hideout/service/reaction/ReactionServiceImpl.kt
@@ -11,7 +11,7 @@ class ReactionServiceImpl(
     private val reactionRepository: ReactionRepository,
     private val apReactionService: APReactionService,
     private val reactionQueryService: ReactionQueryService
-) : IReactionService {
+) : ReactionService {
     override suspend fun receiveReaction(name: String, domain: String, userId: Long, postId: Long) {
         if (reactionQueryService.reactionAlreadyExist(postId, userId, 0).not()) {
             reactionRepository.save(
