--- src/main/kotlin/dev/usbharu/hideout/service/user/IUserAuthService.kt
@@ -1,13 +1,0 @@
-package dev.usbharu.hideout.service.user
-
-import java.security.KeyPair
-
-interface IUserAuthService {
-    fun hash(password: String): String
-
-    suspend fun usernameAlreadyUse(username: String): Boolean
-
-    suspend fun generateKeyPair(): KeyPair
-
-    suspend fun verifyAccount(username: String, password: String): Boolean
-}
