+++ src/main/kotlin/dev/usbharu/hideout/service/user/UserAuthService.kt
@@ -1,53 +1,13 @@
 package dev.usbharu.hideout.service.user
 
-import dev.usbharu.hideout.config.Config
-import dev.usbharu.hideout.query.UserQueryService
-import io.ktor.util.*
-import org.koin.core.annotation.Single
-import java.security.*
-import java.util.*
+import java.security.KeyPair
 
-@Single
-class UserAuthService(
-    val userQueryService: UserQueryService
-) : IUserAuthService {
+interface UserAuthService {
+    fun hash(password: String): String
 
-    override fun hash(password: String): String {
-        val digest = sha256.digest(password.toByteArray(Charsets.UTF_8))
-        return hex(digest)
-    }
+    suspend fun usernameAlreadyUse(username: String): Boolean
 
-    override suspend fun usernameAlreadyUse(username: String): Boolean {
-        userQueryService.findByName(username)
-        return true
-    }
+    suspend fun generateKeyPair(): KeyPair
 
-    override suspend fun verifyAccount(username: String, password: String): Boolean {
-        val userEntity = userQueryService.findByNameAndDomain(username, Config.configData.domain)
-        return userEntity.password == hash(password)
-    }
-
-    override suspend fun generateKeyPair(): KeyPair {
-        val keyPairGenerator = KeyPairGenerator.getInstance("RSA")
-        keyPairGenerator.initialize(keySize)
-        return keyPairGenerator.generateKeyPair()
-    }
-
-    companion object {
-        val sha256: MessageDigest = MessageDigest.getInstance("SHA-256")
-        const val keySize = 2048
-        const val pemSize = 64
-    }
-}
-
-fun PublicKey.toPem(): String {
-    return "-----BEGIN PUBLIC KEY-----\n" +
-        Base64.getEncoder().encodeToString(encoded).chunked(UserAuthService.pemSize).joinToString("\n") +
-        "\n-----END PUBLIC KEY-----\n"
-}
-
-fun PrivateKey.toPem(): String {
-    return "-----BEGIN PRIVATE KEY-----\n" +
-        Base64.getEncoder().encodeToString(encoded).chunked(UserAuthService.pemSize).joinToString("\n") +
-        "\n-----END PRIVATE KEY-----\n"
+    suspend fun verifyAccount(username: String, password: String): Boolean
 }
